requirements.txt
langchain
langchain-community
fastapi
uvicorn
python-dotenv